rename: 자체 퀴즈 생성 컴포넌트 파일 위치 및 `own` -> `user` 변경
